0315/requirements.txt
gradio>=4.0.0
plotly>=5.0.0
yfinance>=0.2.0
hopsworks>=3.0.0
pandas>=1.5.0
numpy>=1.23.0
python-dotenv>=1.0.0
